+++ lib/glimmer/dsl/opal/dsl.rb
@@ -24,6 +24,7 @@ require 'glimmer/dsl/opal/rgba_expression'
 require 'glimmer/dsl/opal/custom_widget_expression'
 require 'glimmer/dsl/opal/swt_expression'
 require 'glimmer/dsl/opal/radio_group_selection_data_binding_expression'
+require 'glimmer/dsl/opal/checkbox_group_selection_data_binding_expression'
 
 module Glimmer
   module DSL
@@ -37,6 +38,7 @@ module Glimmer
          combo_selection_data_binding
          list_selection_data_binding
          radio_group_selection_data_binding
+         checkbox_group_selection_data_binding
          data_binding
          font
          layout